+++ apis/team/v1alpha1/zz_teamsettings_types.go
@@ -1,7 +1,3 @@
-// SPDX-FileCopyrightText: 2023 The Crossplane Authors <https://crossplane.io>
-//
-// SPDX-License-Identifier: Apache-2.0
-
 /*
 Copyright 2022 Upbound Inc.
 */
@@ -25,7 +21,7 @@ type ReviewRequestDelegationInitParameters struct {
 
 	// The number of team members to assign to a pull request
 	// The number of team members to assign to a pull request.
-	MemberCount *float64 `json:"memberCount,omitempty" tf:"member_count,omitempty"`
+	MemberCount *int64 `json:"memberCount,omitempty" tf:"member_count,omitempty"`
 
 	// whether to notify the entire team when at least one member is also assigned to the pull request
 	// whether to notify the entire team when at least one member is also assigned to the pull request.
@@ -40,7 +36,7 @@ type ReviewRequestDelegationObservation struct {
 
 	// The number of team members to assign to a pull request
 	// The number of team members to assign to a pull request.
-	MemberCount *float64 `json:"memberCount,omitempty" tf:"member_count,omitempty"`
+	MemberCount *int64 `json:"memberCount,omitempty" tf:"member_count,omitempty"`
 
 	// whether to notify the entire team when at least one member is also assigned to the pull request
 	// whether to notify the entire team when at least one member is also assigned to the pull request.
@@ -57,7 +53,7 @@ type ReviewRequestDelegationParameters struct {
 	// The number of team members to assign to a pull request
 	// The number of team members to assign to a pull request.
 	// +kubebuilder:validation:Optional
-	MemberCount *float64 `json:"memberCount,omitempty" tf:"member_count,omitempty"`
+	MemberCount *int64 `json:"memberCount,omitempty" tf:"member_count,omitempty"`
 
 	// whether to notify the entire team when at least one member is also assigned to the pull request
 	// whether to notify the entire team when at least one member is also assigned to the pull request.
@@ -153,8 +149,8 @@ type TeamSettingsStatus struct {
 // +kubebuilder:storageversion
 
 // TeamSettings is the Schema for the TeamSettingss API. Manages the team settings (in particular the request review delegation settings)
-// +kubebuilder:printcolumn:name="READY",type="string",JSONPath=".status.conditions[?(@.type=='Ready')].status"
 // +kubebuilder:printcolumn:name="SYNCED",type="string",JSONPath=".status.conditions[?(@.type=='Synced')].status"
+// +kubebuilder:printcolumn:name="READY",type="string",JSONPath=".status.conditions[?(@.type=='Ready')].status"
 // +kubebuilder:printcolumn:name="EXTERNAL-NAME",type="string",JSONPath=".metadata.annotations.crossplane\\.io/external-name"
 // +kubebuilder:printcolumn:name="AGE",type="date",JSONPath=".metadata.creationTimestamp"
 // +kubebuilder:resource:scope=Cluster,categories={crossplane,managed,github}
